Code structure. Clean config.py, enviroment path.
