+++ .gitignore
@@ -160,4 +160,5 @@ venv.bak/
 .idea
 .lock
 Pipfile.lock
+app/data/
 data/
