--- docker-compose.yml
@@ -1,29 +1,0 @@
-# this docker-compose.yml is for developmental purposes only
-version: '3.1'
-services:
-  db:
-    container_name: cryptic_dev_mysql_1
-    image: mysql
-    restart: always
-    ports:
-        - "3306:3306"
-    environment:
-        MYSQL_ROOT_HOST: '%'
-        MYSQL_USER: cryptic
-        MYSQL_PASSWORD: cryptic  # TODO pick a secure password
-        MYSQL_DATABASE: cryptic
-        MYSQL_RANDOM_ROOT_PASSWORD: 1
-  s_user:
-    container_name: cryptic_dev_currency_1
-    restart: always
-    build: ./app
-    ports:
-        - "1242:80"
-    depends_on:
-        - db
-    environment:
-        MYSQL_HOSTNAME: "db"
-        MYSQL_DATABASE: cryptic
-        MYSQL_USERNAME: cryptic
-        MYSQL_PASSWORD: cryptic  # TODO adjust this line
-        DEBUG: 1
